requirements.txt
Digikam_DB==0.1.3
matplotlib==3.5.3
networkx==3.1
numpy==1.22.3
pandas==1.4.3
python-dotenv==1.0.0
pyvis==0.3.2
